showCoverageOnHover can be read from a vector layer
This change allows smk to read a "showCoverageOnHover" value from a vector layer's "layer" element in smk-config.json. This change is a companion to a change in smk-cli that will add a UI control to turn the functionality on or off for a vector layer.

showCoverageOnHover can currently be read from the "viewer" element of smk-config.json. This allows a true value in the viewer to override a value in the layer.

If no value is set in a vector layer or in the viewer, the default value of 'false' is used.