Geomark tool added; can draw polygons to layer
This adds support for a new "Geomark" tool (which will eventually be enabled/disabled in smk-cli). Its implementation has been modeled on that of the Version tool.

This also adds the ability to draw polygons and add them to a layer. This is implemented using the Leaflet Draw library, with new files copied from the latest version:  https://github.com/Leaflet/Leaflet.draw/archive/refs/tags/v0.4.14.zip.