Geomark tool added; can draw polygons to layer (#125)
* Geomark tool added; can draw polygons to layer

This adds support for a new "Geomark" tool (which will eventually be enabled/disabled in smk-cli). Its implementation has been modeled on that of the Version tool.

This also adds the ability to draw polygons and add them to a layer. This is implemented using the Leaflet-Geoman library (https://geoman.io/leaflet-geoman).