Leaflet PM lib removed (#130)
This removes the Leaflet PM library, used by the Markup tool. Its functionality now comes from the Leaflet-Geoman library, which was added in an earlier commit and is a drop-in replacement for Leaflet PM.

Options for adding UI controls in the toolbar have been updated. Many of these are the same as what was available in Leaflet PM, but some have also been added in Leaflet-Geoman. Options that set a default value have been removed. New options that were not available or enabled earlier are disabled. As a result, we have the same six controls in the toolbar as before.