Close-alert error fixed
This addresses the error (seen in console logs) given when attempting to call the "handleAlert" function when it is undefined. This adds a check that the function is defined before calling it.

To reproduce the error, draw a geomark shape, save the drawing, and click 'OK' on the alert that appears afterwards.