Removing Geomark "Load File" functionality (#163)
This removes the "Load File" button and related functionality from the Geomark tool. The link to create a geomark from a file must be updated here whenever the link changes, and we have not heard of customers still wanting this functionality (which was carried over from DMF2), so we are removing it.