Commenting out code that incorrectly references other tools
This comments out code that references other tools by using invalid tool names. Tool names were once all lowercase but are now camel case, but some references to the old tool names remain. This comments out such code because it is currently not functional. This may help prevent bugs, and by preserving the code we keep its intent should we want to bring back functionality later.

This also comments out nonfunctional buttons in the Search tool. The Identify and Measure buttons do not have implementation code, and the code called by the Directions button results in a console error because of an undefined value.