Version incremented, redundant font-family style removed (#179)
This increments the version to synchronize with related changes in SMK-CLI where the BC Sans font will be added and set in the main style sheet for created SMK apps. SMK-CLI changes will depend on this incremented version.

This also removes the font-family style in the map frame styling, which has the same value as the font-family style in style.css (added to new applications by 'smk create'). We remove it because it is redundant and because it overrides the font-family change to BC Sans we want to make in smk-cli to style.css.