Geomarks are toggled by pane visibility
Geomarks are currently toggled on and off by adding and removing them from the map. Keeping a full reference to a geomark layer in memory appears to be causing an issue drawing vector basemap layers.

This addresses the issue by storing an id reference to a geomark layer instead of the layer itself. We no longer use the id of the geomark because, when creating a new geomark polygon/multipolygon, we need a reference to bind to the geomark after it's been created. For this purpose, we use a UUID.

Instead of adding and removing geomark layers from the map, we create a custom pane for each geomark layer and add polygon/multipolygons to the custom pane. Toggling then changes the visibility of a custom pane.

This also changes some variable and function names to be clearer.