Can hide Esri feature layer legend
Esri feature layers can have multiple legends to reflect differing symbology that is based on differing values for a field. When the "where" option is used to omit features based on such a field value, the legend for the omitted value still appears.

This change introduces a 'hideLegend' parameter to Esri feature layer configuration. When the names of one or more legend values are used, those legends are omitted from viewer display context and do not display in UI legends.