Improve ordering of layers
This improves the ordering of layers as they display when an SMK app is first loaded. The layer order follows the order of the 'display' section of SMK config for the layers tool EXCEPT GeoJSON/'vector' layers still appear on top.

This adds a 'layers' pane with a z-index of 300 to avoid conflict with other standard layers used by Leaflet. The zIndex assigned from visible layers in viewer.js is corrected to increase on iteration. Code for Leaflet layers is updated to create a new pane for each layer and set each layer's style on the zIndex from view.js.