Improve ordering of layers
This improves the ordering of layers as they display when an SMK app is first loaded. The layer order follows the order of the 'display' section of SMK config for the layers tool EXCEPT GeoJSON/'vector' layers, which continue to appear on top.

The zIndex assigned from visible layers in viewer.js is corrected to increase on iteration. Code for Leaflet layers is updated to create a new pane for each layer, add it to the default pane ('tile' or 'overlay') and set each pane's style to have the zIndex from viewer.js. Earlier code which sets the z-index style on the layer (much of which didn't appear to work properly) is removed.