Add range and marker to Directions tool
This adds a UI control to direction options to allow input of a range as a positive integer. When a range value exists, the route planner now captures the point along the route where the range limit is met. The directions tool now shows a new marker at the range limit point.