Set pane in vector feature path options to improve ordering
This improves ordering between layers of type vector/GeoJSON. Multiple GeoJSON layers have features rendered in a shared svg element inside the overlay pane. This amends the path options of features to specify the pane created for the layer. This results in an svg element in each GeoJSON layer pane.