Vector layer can hide layer symbology when conditional styling used
In some uses of conditional styling, styling is applied to feature values in a way that each feature is included and matches one style, similar to a pie chart. In such situations, including a style for the entire layer in a legend is unnecessary.

This adds a "hideLayer" property to legend configuration in vector layers and to legend display HTML. When added to a vector/GeoJSON layer's legend configuration and set to true, the symbology for the layer (but not its conditional styling) is omitted in the legends.