Updates to legend display when conditional styling used in vector layer
In some uses of conditional styling, styling is applied to feature values in a way that each feature is included and matches one style, similar to a pie chart. In such situations, including a legend with a swatch using the default style is unnecessary.

This change excludes a swatch for the legend having the layer name when conditional styling is used in vector/GeoJSON layers, and an "Other" legend having a swatch using the layer style. This also adds an "excludeOtherLegendWithDefaultStyling" property to legend configuration. When added to a layer's legend configuration and set to true, the "Other" legend is omitted.

* Documentation is updated for vector/GeoJSON layer to include 'legend' and 'excludeOtherLegendWithDefaultStyling'
* Legend swatches are now always omitted for the legend having the layer name when conditional styling is used
* Property "excludeOtherLegendWithDefaultStyling" is added to legend configuration in the vector layer. When set to true, the "Other" legend is omitted.
* Legend tool CSS is updated to add padding