Reinstate vector onEachFeature function
This restores the onEachFeature function to vector layer options for internal layers. It was removed in SMK 1.1.6 when changes were made to support conditional styling. When a radius is used with the Identify tool, its styling is lost and the radius draws as a black circle; this change restores prior radius styling.